(original rev 1488) resolve ambiguity w/ type size
git-svn-id: file:///tmp/current/omni/branches/staging@1257 2ffc8166-1744-45da-a82a-d6ea10f7896b

+++ omni/src/utility/omDataArchiveBoost.cpp
@@ -2,7 +2,7 @@
 
 QDataStream &operator<<(QDataStream & out, const OmIds & set )
 {
-	out << set.size();
+	out << (quint32)set.size();
 
 	OmIds::const_iterator iter;
 	for( iter = set.begin(); iter != set.end(); ++iter ){
